refactor: enshrine note randomness
This PR makes Aztec-nr handle note randomness as a built-in feature.

Before this PR, randomness injection was a userland concern: users of Aztec-nr were expected to remember to include a randomness attribute in the body of their custom notes to protect against pre-image attacks.

After this PR, the randomness is part of the note metadata Aztec-nr handles natively (similar to what happens with storage slot, for example).

Co-authored-by: benesjan <janbenes1234@gmail.com>